pyproject.toml
[tool.poetry]
name = "cdk-lambda-dynamodb"
version = "0.1.0"
description = "AWS CDK Python infrastructure template with Lambda and DynamoDB"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "cdk" },
    { include = "service" },
]

# Lambda runtime dependencies - exported to layer via `poetry export`
[tool.poetry.dependencies]
python = "^3.13"
aws-lambda-powertools = "^3.0.0"
boto3 = "^1.35.0"

# Dev/CDK dependencies - not deployed to Lambda
[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "^2.164.0"
aws-cdk-aws-lambda-python-alpha = "^2.164.0a0"
constructs = "^10.4.2"
cdk-nag = "^2.29.17"
flake8 = "^7.1.1"
black = "^24.10.0"
pylint = "^3.3.1"
mypy = "^1.13.0"
bandit = "^1.7.10"
isort = "^5.13.2"
autoflake = "^2.3.1"
autopep8 = "^2.3.1"
pytest = "^8.3.3"
pre-commit = "^4.0.1"
ruff = "^0.7.3"
syrupy = "^4.8.0"
boto3-stubs = "^1.35.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration
[tool.ruff]
target-version = "py313"
line-length = 120
exclude = [".venv", "cdk.out", ".build"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
]
ignore = [
    "S101",   # assert allowed in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", ".venv", "cdk.out", ".build"]
skips = ["B101"]  # assert_used

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

# MyPy configuration
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = [".venv", "cdk.out", ".build"]
